backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG Chatbot Backend for Physical AI & Humanoid Robotics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "selenium>=4.39.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]
